fixes for default off FF, overflow, code split
