+++ web/packages/studio/src/components/common/SelectableCard/index.tsx
@@ -90,17 +90,17 @@ export const SelectableCard: FC<SelectableCardProps> = ({
     disabled={disabled}
     aria-pressed={selected}
     aria-label={ariaLabel}
-    title={disabled ? disabledReason : undefined}
-    className={`flex w-[240px] justify-between flex-col items-start gap-1.5 rounded-md border bg-surface-raised px-2 py-1.5 text-left transition-colors ${
+    title={disabled ? disabledReason : title}
+    className={`flex w-[240px] justify-between flex-col items-start gap-1.5 overflow-hidden rounded-md border bg-surface-raised px-2 py-1.5 text-left transition-colors ${
       disabled
         ? 'cursor-not-allowed opacity-50 border-base'
         : `cursor-pointer hover:border-strong hover:bg-surface-hover focus-visible:border-strong focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-(--color-brand,#76b900) ${selected ? 'border-strong' : 'border-base'}`
     } ${className ?? ''}`}
   >
-    <Stack gap="1.5">
-      <Flex className="w-full items-center gap-2">
+    <Stack gap="1.5" className="w-full min-w-0">
+      <Flex className="w-full min-w-0 items-center gap-2">
         {leading}
-        <Stack gap="density-xxs" className="min-w-0">
+        <Stack gap="density-xxs" className="min-w-0 flex-1">
           <Text kind="body/semibold/sm" className="truncate text-primary">
             {title}
           </Text>
